Do not group highlights by field and stop trying to find a field label
Depending on the index configuration, it can be impossible to find the
correct label for a particular Solr field. For instance if a Solr field
is used by more than one "search field", or if more than one Omeka
properties are indexed in the same Solr field.

Also there was some flaws in the current implementation:
* if a corresponding "search field" was not found, an entire resource
  was loaded for every label we try to find, which could have a negative
  performance impact
* searching for a mapping would not work if the "source" was something
  other than a resource's own property (eg. resource_class or
  media/dcterms:title)

Because of the complexity of it, it's better to not try to find a label
for a Solr field, at least for now. It can be added later.

Other changes:
* Remove the code related to building the position string ("i/total").
  It's only for display purpose and should be done in the theme or a
  template file in module Search