Stop using PHP Solr extension
This extension is never installed by default so it requires an
additional step in the install process, and what it provides can easily
be reimplemented without any additional dependency.

Configuration has changed and now we only need to give the Solr URI
(instead of hostname, port, and path). This means we can now use HTTPS
(useful when Solr is not on the same network as Omeka).

The minimum version of Solr is set to 5.1.0 (which introduces the JSON
Request API)